Refactor Agent and SearchState using State superclass
Because a lot of similar functionality to keep track of states was needed
in both the Agent class and the SearchState class, this functionality
was implemented in a separate class State. Agent and SearchState was
then changed to inherit from this class, so that it can be implemented
only one time.