Fix agent not being able to open doors
Fixed a bug where the agent was not able to open doors. The reason was
that the state of the map itself is no checked when using the
sameState() method to check if two states are repeats. This means that
after using a key the agent is seemingly in the same state, so this
action is pruned off as a repeated state.

Because checking the entire map is very expensive, some variables
were introduced to keep count of how many:
- Doors have been opened
- Trees have been cut down
- Tiles have been blown up

This is used as a proxy to be able to easily judge if the map has
changed or not