Keep track of which tiles changed, not how many
Using variables to track the number of tiles that have been opened, cut
or blown up led to different states being impossible to tell apart.
When removing duplicate states this is a problem, because states that
are not duplicates are removed. By keeping track of the tiles themselves
and not the number, this should be solved.