Fix hash collisions in State.hashCode
Hashing of the linked lists of tiles summed the x and y coordinate before
multiplying. This meant that for tiles with x1 = y2 and y1 = x2 the
sum added to the hash was identical. For states where this was the only
difference this created a collision, and states were wrongly removed
from open states when searching.